# CacheKos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Lorem Ipsum",
        "contact": {
            "email": "hi.wasissubekti02@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get all payments",
                "description": "Retrieve all payments data",
                "operationId": "a67c4dcf277fa699515fd3b6d4835388",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all payments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data payments berhasil diambil"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "invoice": {
                                                        "type": "string"
                                                    },
                                                    "jumlah_bayar": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "tanggal": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "id_resident": {
                                                        "type": "integer"
                                                    },
                                                    "keterangan": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "kurang_bayar": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "nullable": true
                                                    },
                                                    "grand_total": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "bukti_bayar": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "created_by": {
                                                        "type": "integer"
                                                    },
                                                    "updated_by": {
                                                        "type": "integer",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new payment",
                "description": "Add a new payment record",
                "operationId": "92ea432a93ffed2bbf16f721fdee61e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoice",
                                    "jumlah_bayar",
                                    "status",
                                    "tanggal",
                                    "id_resident",
                                    "grand_total",
                                    "created_by"
                                ],
                                "properties": {
                                    "invoice": {
                                        "type": "string",
                                        "example": "INV001"
                                    },
                                    "jumlah_bayar": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Paid"
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-06"
                                    },
                                    "id_resident": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Payment for services",
                                        "nullable": true
                                    },
                                    "kurang_bayar": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0,
                                        "nullable": true
                                    },
                                    "grand_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "bukti_bayar": {
                                        "type": "string",
                                        "example": "file_url",
                                        "nullable": true
                                    },
                                    "created_by": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "updated_by": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data payments berhasil ditambahkan"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment by ID",
                "description": "Retrieve a specific payment by its ID",
                "operationId": "055c1b4080a8c5b59d53aa69a505abf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved payment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data payment berhasil dimuat."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data payment tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Update an existing payment",
                "description": "Update a payment record by ID",
                "operationId": "256d3cd03413b7fd8f82cf4491f08847",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoice",
                                    "jumlah_bayar",
                                    "status",
                                    "tanggal",
                                    "kurang_bayar",
                                    "grand_total"
                                ],
                                "properties": {
                                    "invoice": {
                                        "type": "string",
                                        "example": "INV001"
                                    },
                                    "jumlah_bayar": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Paid"
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-06"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Payment for services",
                                        "nullable": true
                                    },
                                    "kurang_bayar": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0
                                    },
                                    "grand_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "bukti_bayar": {
                                        "type": "string",
                                        "example": "file_url",
                                        "nullable": true
                                    },
                                    "updated_by": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Delete a payment",
                "description": "Delete a payment record by ID",
                "operationId": "fd511b89a9c38f94cb4d1e5a35fe9f40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted payment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data payments berhasil dihapus"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pembayaran tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Payments",
            "description": "Payments"
        }
    ]
}